% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/help_buttom.R
\name{help_buttom}
\alias{help_buttom}
\title{Add a question mark helper button}
\usage{
help_buttom(local, body, title = "Ajuda")
}
\arguments{
\item{local}{A shiny input element.}

\item{body}{Text to include in the body.}

\item{title}{Text to include in the title.}
}
\description{
Shortcut to add an action button with a question mark help icon. USE INTERNAL ONLY.
}
\note{
Use internal only.
}
\keyword{internal}
